pyproject.toml
[project]
name = "2a2020600060-hosyminhha-day12"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.14.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.136.0",
    "google-generativeai>=0.8.6",
    "openai>=2.32.0",
    "psutil>=7.2.2",
    "pydantic>=2.13.1",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "uvicorn[standard]>=0.44.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
# "**/{tests,docs,tools}/*" = ["E402"]
"**/tests/*" = ["S101"]

[tool.ruff.format]
# 5. Use single quotes in `ruff format`.
quote-style = "single"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 72

[tool.ruff.lint]
# select = [
#     "ALL",
# ]
select = [
    "F",      # pyflakes (runtime errors, unused imports/variables)
    "E", "W", # pycodestyle (style and whitespace)
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade (modernize syntax)
    "B",      # flake8-bugbear (common bugs & best practices)
    "SIM",    # flake8-simplify (code simplification)
    "C4",     # flake8-comprehensions (list/set/dict comprehension help)
    "TC",     # flake8-type-checking (move typing imports to `if TYPE_CHECKING`)
    "PL",     # flake8-unused-arguments
    "RET",    # flake8-return (return statement issues)
    "ARG",    # pylint rules (a subset)
    "ANN",    # type annotations enforcement
    "COM",    # flake8-commas
    "C90",    # maccabe complexity
    "D",      # pydocstyle
    "A",      # flake8-builtins

    # "YTT",
    # "DTZ",
    # "ISC",
    # "ICN",
    # "INP",
    # "PIE",
    # "PT",
    # "TID",
    # "PLC",
    # "PLE",
    # "PLR",
    # "PLW",
    # "RUF",
    # "PYI",
]

ignore = [
    'Q000',   # Double quote
    'ERA001', # Code comment
    'E501',   # line too long
    'F841',   # pyright already supports this
    'F401',   
    'D416',   # numpydocstyle does not use colon in section name
    'D100', 'D101', 'D102', 'D103', 'D104', 'D107', # Missing docstrings
    'ANN001', 'ANN201', 'ANN202', # Missing type annotations
    'D205', 'ARG001', 'PLW0603', 'PLR2004', 'COM812', # Other common violations in this codebase
]

fixable = ['ALL']
unfixable = []

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.11",
]
